fix for Solaris - must cast void* to long before int - works on x86_64 and x86
git-svn-id: http://svn.code.sf.net/p/nitro-nitf/code/trunk@824 7dc3ea37-dbd3-4c13-bb9a-7ccd57e75167

+++ c++/nitf/include/nitf/IOHandle.hpp
@@ -76,7 +76,7 @@ public:
     //!  Get native object
     nitf_IOHandle getHandle()
     {
-        return (nitf_IOHandle) getNativeOrThrow()->data;
+        return (nitf_IOHandle)(long) getNativeOrThrow()->data;
     }
 
     void create(const std::string& fname, nitf::AccessFlags access =
